Add models for Api and Cli
